一旦astro5のファイルで更新
