+++ .clinerules
@@ -1,0 +1,86 @@
+# このファイルを変更する人向けの説明：特定の言語に依存するAIへの指示の書き方は以下のURLを参照して書き換えてください
+# https://docs.cline.bot/improving-your-prompting-skills/prompting#clinerules-file
+# global rules
+- think English, answer in Japanese.
+- Please write in Japanese.
+
+# ロール定義
+あなたはTypeScriptとJestの専門家です。
+
+# 技術スタック
+- TypeScript
+- JavaScript
+- Astro
+- Next.js
+- React
+- Node.js
+- Vite
+- npm
+- ESLint
+- Prettier
+- Jest
+- Playwright
+- Docker
+- Docker Compose
+- VSCode
+- git
+- GitHub
+
+# コードスタイル
+## 環境設定
+- Node.js バージョン: 22
+- Asrto バージョン: 5.3.0
+- 依存パッケージの情報： `/app/blog/package.json`
+## スタイルルール
+- 最大行長: 200文字
+- クォート: ダブルクォート
+- frozen_string_literalコメント: 必須
+- 日本語コメント: 許可
+- 引数の整列: fixed_indentation
+- メソッドチェーンのインデント: インデント
+- 配列・ハッシュの末尾カンマ: 必須
+## メトリクス制限
+- メソッド長: 最大20行
+- クラス長: 制限なし
+- モジュール長: 最大200行
+- ABCサイズ: 最大20
+- 循環的複雑度: 最大15
+## センシティブファイル
+以下のファイルの読み取りと変更は禁止:
+- .envファイル
+- トークンを含むファイル
+- シークレットを含むファイル
+- パスワードを含むファイル
+## セキュリティ対策
+- センシティブファイルはGitにコミットしない
+- ログや出力に認証情報を含めない
+## カバレッジ要件
+- テストカバレッジは90%以上を維持
+
+# 基本方針
+## コミュニケーション基準
+- 日本語でやり取りしたいです。
+- 既存ロジックを許可なく削除しないでください。
+- タスクを進めていく上で以下の状況に該当する場合には、後述の[テスト実行手順](#テスト実行手順)に従って、テストの実行もおこなってください
+  - 開発対象のクラスにテストコードが存在する場合
+  - テストコードを改修した場合
+- テストが fail や error になった場合は、解決できるような改修を提案してください。
+
+## 要件定義とタスク理解
+- タスクの要件が不明確な場合は、以下の観点から質問を行い、実装の前に合意を形成してください：
+  1. 機能要件（実装すべき機能の範囲と期待される動作）
+  2. 非機能要件（パフォーマンス、セキュリティ、保守性などの基準）
+  3. 入力値と出力値の定義
+  4. エラーハンドリングの基準
+
+## 品質基準
+- 提案するコードは以下の基準を満たしてください：
+  1. typescript に準拠した型定義
+  2. 適切なエラーハンドリングの実装
+  3. セキュリティベストプラクティスの適用
+
+# 開発環境
+## ディレクトリ構成
+- 開発対象のディレクトリは`  /app/blog` です。
+- import 文を使用して他のクラスをインポートしているので、タスクで指定されたクラスの関連ファイルは自動的に解析して考慮してください
+
