+++ .textlintrc
@@ -1,0 +1,12 @@
+{
+  "filters": {},
+  "rules": {
+    "preset-ja-spacing": {
++      "ja-space-between-half-and-full-width": {
++            "space": "always"
++        }
++    },
+    "preset-ja-technical-writing": true,
+    "spellcheck-tech-word": true
+  }
+}
