--- blog/src/content/blog/aws-0001.md
@@ -1,31 +1,0 @@
----
-title: "LambdaでAWSの料金を毎日Slackに通知した結果課金されてしまった"
-pubDate: "2017-09-04"
-updatedDate: "2022-05-10"
-description: 'LambdaでAWSの料金を毎日Slackに通知した結果課金されてしまった'
-image:
-    url: '/images/header/computer_monku.png' 
-    alt: ''
-tags: ["AWS"]
----
-# はじめに
-請求アラートは設定していますが、それとは別に[AWSの料金をLambdaのcronで定期チェックしてSlackに通知する](http://qiita.com/saku/items/fc6b70a420a5c510de2b)を設定して１ヶ月間運用してみました。  
-毎日金額がわかるとやっぱり安心できますよね。
-
-# 推移
-
-![](/images/aws-0001/2017-09-04-4.png)
-
-![](/images/aws-0001/2017-09-04-5.png)
-
-![](/images/aws-0001/2017-09-04-6.png)
-
-![](/images/aws-0001/2017-09-04-7.png)
-
-![](/images/aws-0001/2017-09-04-8.png)
-
-![](/images/aws-0001/2017-09-04-9.png)
-
-# おわりに
-安心のために 2 円(2017/09/01 のレート)支払うことになった。
-悲しい。
