--- blog/src/content/blog/covid-19.md
@@ -1,78 +1,0 @@
----
-title: "Covid-19感染体験記"
-pubDate: "2022-08-04"
-updatedDate: "2022-08-04"
-description: ''
-image:
-    url: '/images/header/medical_vaccine_covid19.png' 
-    alt: ''
-tags: ["日常"]
----
-# 概要
-今流行のコロナに罹ってしまったので症状などを残しておくものです。
-
-
-## 時系列
-* 7/18(月) -8 日目海の日。近所のスーパー銭湯でお風呂。サウナの人数制限が無くなっていて、緩和されたなぁと思った。
-* 7/19(火) -7 日目在宅勤務。終日自宅。
-* 7/20(水) -6 日目出社。1 人で外食して帰宅。
-* 7/21(木) -5 日目在宅勤務。終日自宅。
-* 7/22(金) -4 日目出社。1 人で外食して帰宅。
-* 7/23(土) -3 日目会社の同僚と 2 時間程度対面で食事。その後同僚のコロナ陽性は無し。
-* 7/24(日) -2 日目休みの日なので引き籠もる。終日自宅。
-* 7/25(月) -1 日目特に体調異変なし。会社終わりに同僚と食事。<br/>
-  マスク無しで 1 時間ほど対面で食事。その後同僚のコロナ陽性は無し。
-* 7/26(火) 0 日目今考えるとちょっと体温が高く感じた。平熱+0.5℃ぐらい。<br/>
-  あと心拍数が平常時+10 程度上がっていた。<br/>
-  その日の夜発熱。翌日の快方を祈りながら解熱剤を飲んで寝るも喉が痛くて 1 時間ごとに起きる。<br/>
-  インフルエンザ時と同様に悪寒あり寒いため秋冬用の毛布使用。<br/>
-* 7/27(水) 1 日目 38.1℃。<br/>
-  行政のフロー通りかかりつけ医に連絡するも発熱外来やっていないということで発熱外来に電話。<br/>
-  16:30 に予約。行くころには 38.7℃。<br/>
-  発熱外来ではカーテンで仕切られた(といっても上と下空いてるが…)ところに個別に通される。<br/>
-  PCR 検査は一般的な試験管にストローで唾液を溜めるタイプ。<br />
-  周りの問診が聞こえるが自分と同じ症状でコロナだなーと思った。<br/>
-  5 日分の薬（解熱剤、喉の薬、総合かぜ薬）を出されて帰宅。<br/>
-  喉の激痛でゼリー飲料ぐらいしか受け付けない。一番つらかった。<br/>
-  前日と同様に悪寒があり秋冬用の毛布使用。 
-* 7/28(木) 2 日目 38.9℃。<br/>
-  のどの痛みは継続。<br/>
-  無限にタンが出てきてタンとつばで溺れる夢見る。苦しい。<br/>
-  昼過ぎに知らせが来るとのことだった PCR 検査の結果は 15:30 ごろ。病院も大変そう。<br/>
-  陽性判定とのことで会社に連絡や濃厚接触者へ連絡。
-* 7/29(金) 3 日目 38.0℃。<br />
-  ちょっと会社のチャットが気になって、ちょっと見てみたけどすぐ疲れて駄目だった。<br/>
-  のどの痛み減ってきておかゆ食べられるように。<br/>
-  朝は食べた後気持ち悪くなるも夜は大丈夫。<br/>
-  同居家族全員が濃厚接触者となり外出が出来ないため、姉が物資を持ってきてくれる。<br/>
-  解熱剤など薬が切れかけていたので嬉しい。<br/>
-  アクエリアスばっかり飲んでいたからか歯が痛い。と思ったらコロナからの攻撃だったよう。<br/>
-  咳が出るようになる。<br/>
-  夜は首のリンパのあたりが痛む横向いて睡眠出来ず仰向けで睡眠を取ると痰が気管を塞ぎ 1 時間で目が覚める。
-* 7/30(土) 4 日目 37.0℃ 鼻水タンくしゃみせき
-* 7/31(日) 5 日目 36.5℃ 鼻水タンくしゃみせき<br/>
-  ようやく保健所から連絡あり。自宅療養を続けるように指示される。
-* 8/1(月) 6 日目 36.5℃ タンくしゃみせき
-* 8/2(火) 7 日目 36.8℃ タンくしゃみせき
-* 8/3(水) 8 日目 36.9℃ タンくしゃみせき
-* 8/4(木) 9 日目 36.3℃ タンくしゃみせき
-* 8/5(金) 10 日目 36.0℃ タンくしゃみせき<br/>
-  保健所から自宅療養期間が終了し外出してもよいとの連絡。
-
-## 感染の心当たり
-実際全然無くて、発病前に食事を共にした同僚はどちらも陰性で症状なし。<br/>
-それ以外は会社と自宅の往復で一人外食と電車ぐらい。<br/>
-海の日のスーパー銭湯はちょっと遠いけどコロナの潜伏期間はちょっとググった感じ 1～10 日（平均 2～3 日）とあったのであり得ないことではないのかな。
-
-## コロナ自宅療養であって助かったもの
-* ウィダーインゼリー
-* 大量の箱ティッシュの買い置き
-* 氷枕
-* 咳止めトローチ
-* レトルトのおかゆ
-* 喉スプレー
-
-## おわりに
-幸運なことに軽症と言われるぐらいの症状でしたがインフルエンザよりも辛いので罹らないなら罹らないほうがいい。
-
-思い返すと痛いところが徐々に移動するなど嫌らしいウィルス。
