--- blog/src/content/blog/development-0001.md
@@ -1,104 +1,0 @@
----
-title: "個人開発で使っている開発ツールやエディタ紹介"
-pubDate: "2017-08-10"
-updatedDate: "2022-05-10"
-description: ''
-image:
-    url: '/images/header/computer_programming_man.png' 
-    alt: ''
-tags: ["Windows","開発環境"]
----
-
-# 概要
-
-現在個人開発の場で使用しているツールやエディタなどを晒してみます。 下記の目的で使用しています。
-
-- Android アプリ制作
-- Web API 及び Web アプリ制作(Ruby,PHP)
-  ![](/blog/public/images/development-0001/5151011a-2f4b-0b1e-6e84-1576f644d83b.png)![](/blog/public/images/development-0001/57b87757-94c4-f8d5-337b-63fe9565a7b3.jpeg)![](images/development-0001/1388ef62-7dc3-3a01-ab17-c3a1af0f45b8.png)
-
-## 作業環境
-
-### ディスプレイ
-
-![](/blog/public/images/development-0001/872f7dff-049a-3654-7ea9-a952d66ac29d-1024x768.jpeg)
-
-![](/blog/public/images/development-0001/29ac423f-e580-7ae7-27db-508a568a6594-1024x768.jpeg)
-ディスプレイは 3 台使っています。 全てディスプレイアームを使用して高さや角度が簡単に調整できるようにしています。 ディスプレイの上にはキングジムのディスプレイボードを設置して小物を置けるように工夫しています。
-
-### PC
-
-<span class="strike">そろそろ Ryzen にしたいですが</span>[Ryzenにしました！](http://qiita.com/toshi-click/items/40895efc0fd8e7c075a7)こんな感じで作業しています。
-
-#### メインPC
-
-| パーツ    | スペック                            |
-|:-------|:--------------------------------|
-| CPU    | AMD Ryzen7 1700 8コア 16スレッド      |
-| メモリ    | DDR4-2666 32GB                 |
-| ディスク   | NVMe SSD M8Pe PX-512M8PeG 512GB |
-| グラフィック | AMD Radeon R9 270X              |
-| ディスプレイ | WUXGA 23～24インチ 3枚               |
-
-#### サブ Surface Pro 4
-
-| パーツ  | スペック    |
-|:-----|:--------|
-| CPU  | Core i7 |
-| メモリ  | 16 GB   |
-| ディスク | 256 GB  |
-
-### キーボード
-
-- Logicool G710p メカニカルキーボード
-  ![](/blog/public/images/development-0001/1bf97b62-365b-5d0f-38eb-9e16d899f1b4-1024x768.jpeg)
-
-### マウス
-
-- Logicool G700s
-  ![](/blog/public/images/development-0001/cd2fd4fe-7147-c1f8-83af-1a68d77b900e.jpeg)
-
-# バージョン管理
-
-## GitHub
-
-![](/blog/public/images/development-0001/6f7b8440-feff-346b-17cc-ac7e702100fd.png)
-
-## GitLab
-
-![](/blog/public/images/development-0001/37d0f89e-4903-ee06-62e9-d76900b08baa.png)
-
-定番の Gihtub と GitLab。 Github は公開しても問題のないリポジトリ(古くなってもう使ってないとか)を GitLab から移動して公開しています。
-
-GitLab はプライベートリポジトリが無料かつ CI も使えるので Android アプリのコード等を置いています。
-
-## リモート会議
-
-![](/blog/public/images/development-0001/8a16df81-beec-377b-b777-fdf181fcc23f.png)
-Skype を使っています。
-
-## チャットツール
-
-![](/blog/public/images/development-0001/42fffbb6-bf61-83f6-6189-a86fe0778370.png)![](/blog/public/images/development-0001/8a16df81-beec-377b-b777-fdf181fcc23f.png)
-主に Slack を使用していますが、会社では Chatwork を使用しているため両方使用しています。
-
-## エディタ
-
-### IntelliJ IDEA
-
-![](/blog/public/images/development-0001/02f5fee2-5919-e4dc-142c-30f068395b9e.jpeg)
-Android、Ruby、PHP が 1 つの IDE で出来て便利なので有償の [Ultimate版](https://www.jetbrains.com/idea/) を使っています。「コレが無いと開発できない」とまで言われているみたいです。
-
-### Vidual Studio Code
-
-![](/blog/public/images/development-0001/d9592da6-35c4-cd22-98a8-1ce300463cf6.png)
-テキストエディタとして使用しています。 
-
-昔は秀丸を使っていましたが、Markdown のプレビューなど便利なのでこちらに移行しました。
-
-会社ではメインの開発エディタとして使っている人も居るので、JS を書くぐらいだったらこれだけでいいです。
-
-### Android Studio
-
-![](/blog/public/images/development-0001/2a365717-0e24-e6c0-cab6-b1e8cdcdf66f.png)
-IntelliJ IDEA でも出来ますが、とりあえず入れてる感じです。
