--- blog/src/content/blog/subjectnews001.md
@@ -1,48 +1,0 @@
----
-title: "#it業界クソ現場オブザイヤー2018厳選集"
-pubDate: "2018-11-20"
-description: ''
-image:
-    url: '/images/header/internet_enjou_sns_woman.png' 
-    alt: ''
-tags: ['Twitter','ツイッター','トレンド','ハッシュタグ']
----
-https://twitter.com/nekobass9425/status/1064332784506830848
-
-https://twitter.com/tsubaki_t1/status/1063986288561287168
-
-https://twitter.com/collin0525/status/1063470855060021248
-
-https://twitter.com/henkeizyu/status/1063993855505129472
-
-https://twitter.com/popow65535/status/1063436295899234304
-
-https://twitter.com/xSaludSaru/status/1063453244410789889
-
-https://twitter.com/MakotoK/status/1064700714771869696
-
-https://twitter.com/arito_lydian/status/1064704236020105216
-
-https://twitter.com/Canary_Kun/status/1063249930288951296
-
-https://twitter.com/0083DelazFleet/status/1064096159000457216
-
-https://twitter.com/yuna_priest/status/1063278259217940480
-
-https://twitter.com/hidamare/status/1063585162070388736
-
-https://twitter.com/IaCthugha/status/1063423523593498624
-
-https://twitter.com/taka_sagawa/status/1063602887517970432
-
-https://twitter.com/yonemura2006/status/1063247424624320513
-
-https://twitter.com/MUGI1208/status/1063575171515133952
-
-https://twitter.com/naosi/status/1063984511224926208
-
-https://twitter.com/TsuyudakuVanda/status/1064447432061181953
-
-https://twitter.com/sawanow/status/1064504162241540097
-
-https://twitter.com/mamurudesuyo/status/1063768650664960001
