--- blog/src/content/config.ts
@@ -1,24 +1,0 @@
-import { defineCollection, z } from 'astro:content';
-
-const blog = defineCollection({
-  // Type-check frontmatter using a schema
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    // Transform string to Date object
-    pubDate: z
-      .string()
-      .or(z.date())
-      .transform((val) => new Date(val)),
-    updatedDate: z
-      .string()
-      .optional()
-      .transform((str) => (str ? new Date(str) : undefined)),
-    image: z.object({
-      url: z.string().optional(),
-      alt: z.string().optional()
-    }),
-  }),
-});
-
-export const collections = { blog };
