+++ my_secret.txt
@@ -1,0 +1,1 @@
+hogehoge
