+++ prompt
@@ -1,0 +1,44 @@
+# 例
+GoでSNSを開発する手順書を作成してください。
+Markdownで4個の大分類、16個の中分類、64個の小分類に分けて説明してください。
+各分類には番号をふってください。
+小分類の下位にMarkdownの番号リストで手順を記載してください。
+開発とテスト以外の手順を記載しないでください。
+また、下記ライブラリを使用してください。
+* github.com/labstack/echo
+* github.com/mattn/go-sqlite3
+* github.com/joho/godotenv
+
+# 手順書をClineに入力する
+上記で生成した手順書をmdにして保存します
+そのmdを参照することで、このようにClineに渡すプロンプトがものすごく短くなります。
+#### Clineへの指示は以下のように
+@/chart.md の1.の作業を実行してください。
+
+
+
+
+# .clinerulesから一時的に抜いてる設定
+# テスト
+- テストファイルを作成、変更した場合はminitestを実行すること。
+
+## テストケース名の規約
+
+- テスト名は"test #メソッド名"または"test should ..."の形式で記述
+- スネークケースを使用
+- 具体的な機能や期待する動作を明確に表現
+
+## テストファイルの構成
+
+- モデルテストにはsetupメソッドを含める
+- テストケースは機能単位でグループ化
+- フィクスチャファイルには適切なコメントを含める
+# テスト実行手順
+テストの実行には `テストコードのクラスのパス` が与えられている必要があります。
+もしもユーザーからテストコードのクラスのパスが与えられていない場合は質問を行い、パスを特定してください。
+
+## 実行例
+テスト実行コマンド：
+```bash
+npx playwright test index.spec.ts
+```
