+++ testrule
@@ -1,0 +1,38 @@
+# Testing Rules
+
+## 1. テストファイルの基本構造
+- ファイル配置・命名規則（`{ComponentName}.test.tsx`）
+- 用途別基本インポート（通常/GraphQL/カスタムフック）
+- テストファイルテンプレート（describeブロック、setupPage関数）
+
+## 2. モックパターン
+- 自動設定されるモック（SVG）
+- よく使うモックパターン（コンポーネント、カスタムフック、ストレージ、GraphQL）
+- 状態管理モック（jotai）
+
+## 3. 要素の検索優先順位
+1. Role-based → screen.getByRole('button', { name: 'ボタン名' })
+2. テキストベース → screen.getByText('表示テキスト')
+3. TestID（最終手段）→ screen.getByTestId('test-id')
+
+## 4. テストケースパターン
+- 表示・非表示確認（動的な要素、条件による表示変化）
+- ユーザーインタラクション
+- 非同期処理（ローディング、APIレスポンス）
+- ナビゲーション、モーダル
+- 状態管理テスト (ストレージ操作)
+- エラーハンドリング
+
+## 5. テストケース命名規則
+'[条件]のとき、[具体的な期待結果]'
+- 例 'プッシュ通知タップ時、通知データを既読に更新し通知詳細画面に遷移する'
+
+## 6. テスト観点ごとに抑えてほしいポイント(必須・推奨別)
+- ユーザーインタラクション
+- データフェッチ
+- 状態による表示確認
+- フォーム操作
+- データ有無による確認
+- コンポーネント間連携
+- アクセシビリティ
+- 条件分岐による確認パターン (例:条件を満たす場合の挙動/条件を満たさない場合の挙動)
